Move REST API common code to rest_api.cpp/h
