+++ simple_metering.cpp
@@ -1,3 +1,4 @@
+#include <math.h>
 #include "de_web_plugin.h"
 #include "de_web_plugin_private.h"
 #include "device_descriptions.h"
