+++ zcl_tasks.cpp
@@ -10,6 +10,7 @@
 
 #include <QString>
 #include <QVariantMap>
+#include <math.h>
 #include "de_web_plugin.h"
 #include "de_web_plugin_private.h"
 #include "colorspace.h"
